requirements.txt
asyncio==3.4.3
elevenlabs==1.50.5
python-dotenv==1.0.1
websockets==14.2
